+++ ltlcross_wrapper/__init__.py
@@ -2,6 +2,6 @@
 
 __version__ = "0.6.1dev"
 
-from ltlcross_wrapper.analyzer import ResAnalyzer, update_resfile, merge_resfiles, remove_tools
+from ltlcross_wrapper.analyzer import ResAnalyzer, update_resfile, merge_resfiles, remove_tools, gather_cumulative, gather_mins
 from ltlcross_wrapper.modular import Modulizer
 from ltlcross_wrapper.modular import GoalModulizer